Pass handling of library to its own class and Introduced an Scene manager.
All SDL initializations are encapsulated on the SDLEngine class. Some
classes are affected in that they are simpler to initialize (e.g., Text)
as the libraries are loaded at the beggining of the program and thus
there's no need to separate initialization on a class method instead of
the constructor.

The scene manager prototype provides an practical way to change from a
scene. Title screen is still on the prototype stage.